@@ -1,14 +1,17 @@
 {
     "require": {
-        "php": "7.3.*",
+        "php": "7.4.*",
+        "ext-json": "*",
+        "ext-pdo": "*",
+        "ext-fileinfo": "*",
         "slim/slim": "^3.0",
         "ramsey/uuid": "^3.9",
         "robmorgan/phinx": "^0.11.4",
-        "ext-json": "*",
         "nette/database": "^3.0",
-        "ext-pdo": "*",
         "zordius/lightncandy": "^1.2",
-        "mikehaertl/phpwkhtmltopdf": "^2.4"
+        "mikehaertl/phpwkhtmltopdf": "^2.4",
+        "monolog/monolog": "^2.1",
+        "willdurand/negotiation": "^2.3"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.5",
