@@ -4,6 +4,8 @@
 namespace HomeCEU\DTS\Api\DocData;
 
 
+use HomeCEU\DTS\Api\ResponseHelper;
+use HomeCEU\DTS\Entity\DocData;
 use HomeCEU\DTS\Persistence;
 use HomeCEU\DTS\Persistence\DocDataPersistence;
 use HomeCEU\DTS\Repository\DocDataRepository;
@@ -13,18 +15,10 @@ use Slim\Http\Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
 class ListVersions {
-
-  /** @var  Persistence */
-  private $persistence;
-
-  /** @var  DocDataRepository */
-  private $repository;
-
-  /** @var DocDataVersionList */
-  private $useCase;
-
-  /** @var ContainerInterface */
-  private $di;
+  private Persistence $persistence;
+  private DocDataRepository $repository;
+  private DocDataVersionList $useCase;
+  private ContainerInterface $di;
 
   public function __construct(ContainerInterface $diContainer) {
     $this->di = $diContainer;
@@ -34,11 +28,13 @@ class ListVersions {
   }
 
   public function __invoke(Request $request, Response $response, $args) {
-    $versions = $this->useCase->versions($args['docType'], $args['dataKey']);
-    $responseData = [
-        'total' => count($versions),
-        'items' => $versions
-    ];
-    return $response->withJson($responseData);
+      $versions = $this->useCase->versions($args['docType'], $args['dataKey']);
+      $responseData = [
+          'total' => count($versions),
+          'items' => array_map(function (DocData $d) {
+            return ResponseHelper::docDataDetailModel($d);
+          }, $versions)
+      ];
+      return $response->withJson($responseData);
   }
 }
