@@ -8,9 +8,13 @@ use LightnCandy\Flags;
 use LightnCandy\LightnCandy;
 
 class TemplateCompiler {
-  private $flags = Flags::FLAG_HANDLEBARS;
-  private $helpers = [];
-  private $partials = [];
+  private int $flags = Flags::FLAG_HANDLEBARS;
+  private array $helpers = [];
+  private array $partials = [];
+
+  private function __construct() {
+    $this->addHelper(TemplateHelpers::equal());
+  }
 
   public static function create(): self {
     return new self();
@@ -24,20 +28,28 @@ class TemplateCompiler {
     return $this;
   }
 
-  public function addHelper(Helper $helper): void {
+  public function addHelper(Helper $helper): self {
     $this->helpers[$helper->name] = $helper->func;
+    return $this;
   }
 
   public function setPartials(array $partials): self {
     $this->partials = [];
     foreach ($partials as $partial) {
+      $this->checkIsPartial($partial);
       $this->addPartial($partial);
     }
     return $this;
   }
 
-  public function addPartial(Partial $partial): void {
-    $this->partials[$partial->name] = $partial->template;
+  public function addPartial(PartialInterface $partial): self {
+    $this->partials[$partial->getKey()] = $partial->getBody();
+    return $this;
+  }
+
+  public function ignoreMissingPartials(): self {
+    $this->flags |= Flags::FLAG_ERROR_SKIPPARTIAL;
+    return $this;
   }
 
   public function compile(string $template): string {
@@ -49,7 +61,13 @@ class TemplateCompiler {
       ];
       return LightnCandy::compile($template, $options);
     } catch (\Exception $e) {
-      throw new CompilationException("Cannot compile template: {$e->getMessage()}");
+      throw CompilationException::fromException($e);
+    }
+  }
+
+  protected function checkIsPartial($partial): void {
+    if (!($partial instanceof PartialInterface)) {
+      throw new \TypeError('$partials must be an array of PartialInterface objects');
     }
   }
 }
