@@ -1,21 +1,28 @@
 <?php
 
-// this array gets passed into Slim\Container
+namespace HomeCEU\DTS\Api;
 
+use HomeCEU\DTS\Db;
+use HomeCEU\DTS\Db\Config;
+
+// this array gets passed into Slim\Container
 return [
     'settings' => [
         'addContentLengthHeader' => false,
     ],
     'dbConfig' => function ($container) {
-      return \HomeCEU\DTS\Db\Config::fromEnv();
+      return Config::fromEnv();
     },
     'dbConnection' => function($container) {
-      return \HomeCEU\DTS\Db::connection();
+      return Db::connection();
     },
-    'templateCompiler' => function($container) {
-      $compiler = \HomeCEU\DTS\Render\TemplateCompiler::create();
-      $compiler->addHelper(\HomeCEU\DTS\Render\TemplateHelpers::ifComparisonHelper());
-
-      return $compiler;
+    'logger' => function($container) {
+      return Logger::instance();
+    },
+    'errorHandler' => function ($c) {
+      return new ErrorHandler($c);
+    },
+    'phpErrorHandler' => function ($c) {
+      return new ErrorHandler($c);
     }
 ];
