@@ -7,19 +7,18 @@ use HomeCEU\DTS\Db\Config as DbConfig;
 use HomeCEU\DTS\Db\Connection;
 
 class Db {
+  private static Connection $connection;
 
-  /** @var Connection */
-  private static $connection;
-
-  public static function connection() {
-    return self::$connection ?: self::$connection = self::newConnection();
+  public static function connection(): Connection {
+    return self::$connection ?? self::newConnection();
   }
 
-  public static function newConnection(array $options = null) {
-    return Connection::buildFromConfig(static::dbConfig(), $options);
+  public static function newConnection(array $options = null):  Connection {
+    self::$connection = Connection::buildFromConfig(static::dbConfig(), $options);
+    return self::$connection;
   }
 
-  public static function dbConfig() {
+  public static function dbConfig(): DbConfig {
     return DbConfig::fromEnv();
   }
 }
