@@ -5,30 +5,11 @@ namespace HomeCEU\DTS\Entity;
 
 
 use HomeCEU\DTS\AbstractEntity;
-use HomeCEU\DTS\EntityHelperTrait;
 
 class DocData extends AbstractEntity {
-  public $dataId;
-  public $docType;
-  public $dataKey;
-  public $createdAt;
-  public $data;
-
-  use EntityHelperTrait;
-
-  protected static function keys(): array {
-    return [
-        'dataId',
-        'docType',
-        'dataKey',
-        'createdAt',
-        'data'
-    ];
-  }
-
-  public static function fromState(array $state): DocData {
-    $entity = new DocData;
-    self::buildFromState($entity, $state);
-    return $entity;
-  }
+  public string $id;
+  public string $docType;
+  public string $key;
+  public \DateTime $createdAt;
+  public array $data;
 }
