@@ -4,34 +4,13 @@ namespace HomeCEU\DTS\Entity;
 
 
 use HomeCEU\DTS\AbstractEntity;
-use HomeCEU\DTS\EntityHelperTrait;
 
 class Template extends AbstractEntity {
-  public $templateId;
-  public $templateKey;
-  public $docType;
-  public $name;
-  public $author;
-  public $createdAt;
-  public $body;
-
-  use EntityHelperTrait;
-
-  protected static function keys(): array {
-    return [
-        'templateId',
-        'docType',
-        'templateKey',
-        'createdAt',
-        'name',
-        'author',
-        'body'
-    ];
-  }
-
-  public static function fromState(array $state): Template {
-    $entity = new Template;
-    self::buildFromState($entity, $state);
-    return $entity;
-  }
+  public string $id;
+  public string $key;
+  public string $docType;
+  public ?string $name = '';
+  public string $author;
+  public \DateTime $createdAt;
+  public string $body;
 }
