@@ -5,9 +5,10 @@ namespace HomeCEU\DTS;
 
 
 interface Persistence {
-  public function generateId();
-  public function persist($data);
-  public function retrieve($id, array $cols=['*']);
-  public function find(array $filter, array $cols=['*']);
-  public function delete($id);
+  public function generateId(): string;
+  public function persist(array $data): string;
+  public function update(array $data): void;
+  public function retrieve(string $id, array $cols=['*']): array;
+  public function find(array $filter, array $cols=['*']): array;
+  public function delete(string $id);
 }
