@@ -3,16 +3,17 @@
 
 namespace HomeCEU\Tests\Api\DocData;
 
-use HomeCEU\Tests\Api\TestCase;
-use PHPUnit\Framework\Assert;
+use DateTime;
+use HomeCEU\Tests\Api\ApiTestCase;
 
-class ListVersionsTest extends TestCase {
+class ListVersionsTest extends ApiTestCase {
+  const ROUTE = "/api/v1/docdata";
 
   public function testListVersions() {
-    $dataKey = __FUNCTION__; // just in case it doesnt cleanup, you will know where it came from.
-    $this->addFixtureData($dataKey);
+    $key = __FUNCTION__; // just in case it doesnt cleanup, you will know where it came from.
+    $this->addFixtureData($key);
 
-    $uri = "/docdata/{$this->docType}/{$dataKey}/history";
+    $uri = self::ROUTE."/{$this->docType}/{$key}/history";
     $response = $this->get($uri);
     $responseData = json_decode($response->getBody(), true);
 
@@ -21,17 +22,48 @@ class ListVersionsTest extends TestCase {
     $this->assertTotalItems($responseData, 2);
     $this->AssertExpectedVersionItemKeys(
         $responseData,
-        ['dataId', 'docType', 'dataKey', 'createdAt']
+        ['id', 'docType', 'key', 'createdAt']
     );
   }
 
+  public function testResponseFormat() {
+    // load fixture
+    $key = uniqid();
+    $id = uniqid();
+
+    $this->docDataPersistence()->persist([
+        'id' => $id,
+        'docType' => $this->docType,
+        'key' => $key,
+        "createdAt" => new DateTime("2020-10-13 23:47:07"),
+        'data' => ['name'=>'Fred']
+    ]);
+    $expected = [
+        'total' => 1,
+        'items' => [
+            [
+                'id' => $id,
+                'docType' => $this->docType,
+                'key' => $key,
+                "createdAt" => new DateTime("2020-10-13 23:47:07"),
+                "link" => self::ROUTE."/{$id}"
+            ]
+        ]
+    ];
+
+    $uri = self::ROUTE."/{$this->docType}/{$key}/history";
+    $response = $this->get($uri);
+    $responseData = json_decode($response->getBody(), true);
+    $this->assertEquals(json_decode(json_encode($expected), true), $responseData);
+  }
+
   /**
-   * @param string $dataKey
+   * @param string $key
    */
-  private function addFixtureData(string $dataKey): void {
-    $this->addDocDataFixture($dataKey);
+  private function addFixtureData(string $key): void {
+    $this->addDocDataFixture($key);
     $this->addDocDataFixture(uniqid());
-    $this->addDocDataFixture($dataKey);
+    $this->addDocDataFixture($key);
   }
 
   /**
@@ -40,20 +72,18 @@ class ListVersionsTest extends TestCase {
    */
   private function AssertExpectedVersionItemKeys($responseData, array $expectedResponseKeys): void {
     foreach ($expectedResponseKeys as $key) {
-      Assert::assertFalse(empty($responseData['items'][0][$key]));
+      $this->assertFalse(empty($responseData['items'][0][$key]));
     }
 
-    Assert::assertFalse(
-        array_key_exists('data', $responseData['items'][0]),
-        "ERROR: get docdata history should not respond with the data"
-    );
+    $this->assertArrayNotHasKey('data', $responseData['items'][0]);
   }
 
   /**
-   * @param $responseData
+   * @param $responseData array
+   * @param $total int
    */
   private function assertTotalItems($responseData, $total): void {
-    Assert::assertCount($total, $responseData['items']);
-    Assert::assertEquals($total, $responseData['total']);
+    $this->assertCount($total, $responseData['items']);
+    $this->assertEquals($total, $responseData['total']);
   }
 }
