@@ -9,7 +9,7 @@ use HomeCEU\Tests\DTS\TestCase;
 use PHPUnit\Framework\Assert;
 
 class TemplateTest extends TestCase {
-  public function testBuildFromState() {
+  public function testBuildFromState(): void {
     $fakeTemplateArray = $this->fakeTemplateArray();
     $template = Template::fromState($fakeTemplateArray);
     Assert::assertEquals($fakeTemplateArray, $template->toArray());
